feature: goals rework/ implement UI

+++ penny-wise-backend/database/migrations/2025_05_28_101655_add_category_id_to_recurring_transactions_table.php
@@ -9,12 +9,14 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
+    public function up() 
     {
-        Schema::table('recurring_transactions', function (Blueprint $table) {
-            $table->unsignedBigInteger('category_id')->nullable()->after('wallet_id');
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
-        });
+        if (!Schema::hasColumn('recurring_transactions', 'category_id')) {
+            Schema::table('recurring_transactions', function (Blueprint $table) {
+                $table->unsignedBigInteger('category_id')->nullable()->after('wallet_id');
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
+            });
+        }
     }
 
     public function down()
